Refactor migrator
